requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
numpy==1.21.0
pandas==1.3.3
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytest==6.2.5
requests==2.26.0
alembic==1.7.3
prometheus-client==0.11.0
redis==3.5.3
pydantic==1.8.2
aioredis==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
graphene==3.0.0
python-multipart==0.0.5